# feat: reduce GC by not passing the whole struct to compile and process TraceQL metrics