**What this PR does**:
It aims to marginally reduce allocations by not passing the whole struct to compile and process TraceQL metrics.
It also makes the code simpler to understand by not wrapping the start, end, and step, especially for those methods where the request values were being mutated in unclear ways



<details>
  <summary>Query range benchmark results</summary>

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/tempo/tempodb/encoding/vparquet4
                                                                             │   old.txt   │              new.txt               │
                                                                             │   sec/op    │   sec/op     vs base               │
BackendBlockQueryRange/{}_|_rate()/5-11                                         1.284 ± 4%    1.260 ± 3%       ~ (p=0.190 n=10)
BackendBlockQueryRange/{}_|_rate()/7-11                                         1.302 ± 3%    1.293 ± 2%       ~ (p=0.436 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/5-11                               2.388 ± 7%    2.361 ± 6%       ~ (p=0.393 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-11                               2.665 ± 1%    2.635 ± 1%  -1.12% (p=0.043 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/5-11              1.774 ± 8%    1.677 ± 6%       ~ (p=0.075 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-11              2.110 ± 6%    1.971 ± 1%  -6.60% (p=0.000 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/5-11                      2.853 ± 2%    2.836 ± 1%       ~ (p=0.481 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-11                      2.989 ± 1%    2.991 ± 3%       ~ (p=0.971 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/5-11   204.8m ± 0%   204.8m ± 0%       ~ (p=0.481 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-11   205.2m ± 0%   205.9m ± 2%  +0.31% (p=0.004 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/5-11                            183.9m ± 0%   184.2m ± 1%  +0.16% (p=0.009 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/7-11                            184.0m ± 0%   184.4m ± 0%       ~ (p=0.075 n=10)
geomean                                                                        941.2m        927.8m       -1.43%

                                                                             │  old.txt   │               new.txt               │
                                                                             │  MB_IO/op  │  MB_IO/op   vs base                 │
BackendBlockQueryRange/{}_|_rate()/5-11                                        33.74 ± 0%   33.74 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()/7-11                                        33.74 ± 0%   33.74 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(name)/5-11                              42.24 ± 0%   42.24 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-11                              42.24 ± 0%   42.24 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/5-11             34.27 ± 0%   34.27 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-11             34.27 ± 0%   34.27 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/5-11                     37.19 ± 0%   37.19 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-11                     37.19 ± 0%   37.19 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/5-11   34.22 ± 0%   34.22 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-11   34.22 ± 0%   34.22 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{status=error}_|_rate()/5-11                            35.39 ± 0%   35.39 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{status=error}_|_rate()/7-11                            35.39 ± 0%   35.39 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                        36.06        36.06       +0.00%
¹ all samples are equal

                                                                             │   old.txt   │               new.txt                │
                                                                             │  spans/op   │  spans/op    vs base                 │
BackendBlockQueryRange/{}_|_rate()/5-11                                        5.650M ± 0%   5.650M ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()/7-11                                        9.068M ± 0%   9.068M ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(name)/5-11                              5.650M ± 0%   5.650M ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-11                              9.068M ± 0%   9.068M ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/5-11             5.650M ± 0%   5.650M ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-11             9.068M ± 0%   9.068M ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/5-11                     5.650M ± 0%   5.650M ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-11                     9.068M ± 0%   9.068M ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/5-11   197.1k ± 0%   197.1k ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-11   300.6k ± 0%   300.6k ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{status=error}_|_rate()/5-11                            44.67k ± 0%   44.67k ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockQueryRange/{status=error}_|_rate()/7-11                            65.03k ± 0%   65.03k ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                                        1.804M        1.804M       +0.00%
¹ all samples are equal

                                                                             │   old.txt   │              new.txt               │
                                                                             │   spans/s   │   spans/s    vs base               │
BackendBlockQueryRange/{}_|_rate()/5-11                                        4.400M ± 4%   4.485M ± 2%       ~ (p=0.190 n=10)
BackendBlockQueryRange/{}_|_rate()/7-11                                        6.965M ± 2%   7.013M ± 2%       ~ (p=0.436 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/5-11                              2.366M ± 8%   2.393M ± 6%       ~ (p=0.393 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-11                              3.403M ± 1%   3.441M ± 1%  +1.13% (p=0.043 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/5-11             3.186M ± 9%   3.370M ± 6%       ~ (p=0.075 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-11             4.297M ± 6%   4.600M ± 1%  +7.06% (p=0.000 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/5-11                     1.980M ± 2%   1.992M ± 1%       ~ (p=0.481 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-11                     3.034M ± 1%   3.032M ± 3%       ~ (p=0.971 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/5-11   962.3k ± 0%   962.6k ± 0%       ~ (p=0.481 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-11   1.465M ± 0%   1.460M ± 2%  -0.31% (p=0.004 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/5-11                            242.9k ± 0%   242.5k ± 1%  -0.16% (p=0.009 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/7-11                            353.4k ± 0%   352.7k ± 0%       ~ (p=0.075 n=10)
geomean                                                                        1.917M        1.944M       +1.45%

                                                                             │    old.txt    │               new.txt                │
                                                                             │     B/op      │     B/op       vs base               │
BackendBlockQueryRange/{}_|_rate()/5-11                                        51.52Mi ± 47%   39.72Mi ± 53%       ~ (p=0.105 n=10)
BackendBlockQueryRange/{}_|_rate()/7-11                                        47.10Mi ± 27%   49.27Mi ± 31%       ~ (p=0.529 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/5-11                              45.54Mi ± 56%   49.58Mi ± 52%       ~ (p=0.247 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-11                              46.64Mi ± 22%   37.34Mi ± 71%       ~ (p=0.247 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/5-11             43.04Mi ± 42%   46.34Mi ± 72%       ~ (p=0.280 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-11             39.78Mi ± 24%   37.87Mi ± 51%       ~ (p=0.353 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/5-11                     436.0Mi ±  2%   427.3Mi ±  6%       ~ (p=0.393 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-11                     486.4Mi ±  2%   479.8Mi ±  4%       ~ (p=0.123 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/5-11   65.25Mi ±  2%   65.33Mi ±  1%       ~ (p=0.971 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-11   65.09Mi ±  2%   65.63Mi ±  1%       ~ (p=0.105 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/5-11                            5.052Mi ± 23%   5.052Mi ± 11%       ~ (p=0.853 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/7-11                            5.102Mi ± 21%   4.939Mi ± 20%       ~ (p=0.853 n=10)
geomean                                                                        49.28Mi         47.72Mi        -3.15%

                                                                             │   old.txt    │               new.txt                │
                                                                             │  allocs/op   │   allocs/op    vs base               │
BackendBlockQueryRange/{}_|_rate()/5-11                                        55.32k ± 87%   45.39k ±  43%       ~ (p=0.101 n=10)
BackendBlockQueryRange/{}_|_rate()/7-11                                        45.46k ± 99%   45.48k ± 127%       ~ (p=0.780 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/5-11                              60.62k ±  0%   60.67k ± 191%       ~ (p=0.447 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(name)/7-11                              63.83k ±  0%   63.76k ±   0%       ~ (p=0.172 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/5-11             47.54k ±  0%   47.59k ± 444%       ~ (p=0.393 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(resource.service.name)/7-11             47.84k ±  0%   47.83k ±   0%       ~ (p=0.529 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/5-11                     1.072M ±  6%   1.071M ±   7%       ~ (p=0.684 n=10)
BackendBlockQueryRange/{}_|_rate()_by_(span.http.url)/7-11                     1.217M ±  1%   1.222M ±   2%       ~ (p=0.481 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/5-11   346.6k ±  0%   346.6k ±   0%       ~ (p=0.542 n=10)
BackendBlockQueryRange/{resource.service.name=`loki-ingester`}_|_rate()/7-11   346.6k ±  0%   346.7k ±   0%       ~ (p=0.085 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/5-11                            60.30k ±  0%   60.31k ±   2%       ~ (p=0.468 n=10)
BackendBlockQueryRange/{status=error}_|_rate()/7-11                            60.30k ±  0%   60.29k ±   0%       ~ (p=0.926 n=10)
geomean                                                                        123.5k         121.5k         -1.60%
```

</details>


**Checklist**
- [X] Tests updated
- [ ] Documentation added
- [ ] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`